+++ src/eca/server.clj
@@ -45,6 +45,9 @@
 (defmethod lsp.server/receive-request "chat/queryContext" [_ components params]
   (handlers/chat-query-context components params))
 
+(defmethod lsp.server/receive-request "mcp/listServers" [_ components params]
+  (handlers/mcp-list-servers components params))
+
 (defn ^:private monitor-server-logs [log-ch]
   ;; NOTE: if this were moved to `initialize`, after timbre has been configured,
   ;; the server's startup logs and traces would appear in the regular log file