+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Allow comments on `json` configs.
 - Improve MCP tool call feedback.
 - Add support for env vars in mcp configs.
+- Add `mcp/listServers` request.
 
 ## 0.0.4